refactor and fix compat issues for speaker encoder
