+++ TTS/bin/train_encoder.py
@@ -11,10 +11,10 @@ import torch
 from torch.utils.data import DataLoader
 
 from TTS.speaker_encoder.dataset import MyDataset
-from TTS.speaker_encoder.generic_utils import save_best_model
+from TTS.speaker_encoder.utils.generic_utils import save_best_model
 from TTS.speaker_encoder.losses import GE2ELoss, AngleProtoLoss
 from TTS.speaker_encoder.model import SpeakerEncoder
-from TTS.speaker_encoder.visual import plot_embeddings
+from TTS.speaker_encoder.utils.visual import plot_embeddings
 from TTS.tts.datasets.preprocess import load_meta_data
 from TTS.utils.generic_utils import (
     create_experiment_folder, get_git_branch, remove_experiment_folder,
